+++ src/states/controlState.ts
@@ -1,6 +1,11 @@
 import * as ol from 'openlayers';
 import { Time, TimeRange } from '../model/timeSeries';
 import { VIEWER_DEFAULT_API_SERVER } from '../config';
+import { loadUserSettings } from './userSettings';
+
+export type TimeAnimationInterval = 250 | 500 | 1000 | 2500;
+export const TIME_ANIMATION_INTERVALS: TimeAnimationInterval[] = [250, 500, 1000, 2500];
+
 
 export interface ControlState {
     selectedDatasetId: string | null;
@@ -14,7 +19,10 @@ export interface ControlState {
     selectedDrawMode: ol.geom.GeometryType | null;
     timeSeriesUpdateMode: 'add' | 'replace';
     timeAnimationActive: boolean;
-    timeAnimationInterval: number;
+    timeAnimationInterval: TimeAnimationInterval;
+    autoShowTimeSeries: boolean;
+    showTimeSeriesPointsOnly: boolean;
+    showTimeSeriesErrorBars: boolean;
     flyTo: ol.geom.SimpleGeometry | ol.Extent | null;
     activities: { [id: string]: string };
     locale: string;
@@ -22,7 +30,7 @@ export interface ControlState {
 }
 
 export function newControlState(): ControlState {
-    return {
+    const state: ControlState = {
         selectedDatasetId: 'local',
         selectedVariableName: 'conc_chl',
         selectedPlaceGroupIds: null,
@@ -35,14 +43,18 @@ export function newControlState(): ControlState {
         timeSeriesUpdateMode: 'add',
         timeAnimationActive: false,
         timeAnimationInterval: 1000,
+        autoShowTimeSeries: false,
+        showTimeSeriesPointsOnly: false,
+        showTimeSeriesErrorBars: true,
         flyTo: null,
         activities: {},
         locale: 'en',
         dialogOpen: {},
     };
+    return loadUserSettings(state);
 }
 
 // We cannot keep "MAP_OBJECTS" in control state object, because these objects are (1) not serializable
 // and (2) logging actions will cause the browsers to crash
 
-export const MAP_OBJECTS: {[id: string]: ol.Object} = {};
+export const MAP_OBJECTS: { [id: string]: ol.Object } = {};